# swagger.yaml
openapi: 3.0.0
paths:
  /administrators:
    get:
      operationId: AdministratorsController_getContactsV1
      parameters:
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: nextPageToken
          required: false
          in: query
          schema:
            default: null
            type: string
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorsDto"
      security: &a1
        - bearer: []
      summary: Get list of administrators
      tags: &a2
        - Administrators
    post:
      operationId: AdministratorsController_createAdministratorV1
      parameters: []
      requestBody:
        required: true
        description: New administrator DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministratorCreateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorDto"
        "400":
          description: Bad request
          content:
            application/json:
              example: email must be an email
        "406":
          description: Administrator could not be created.
          content:
            application/json:
              example: Administrator with email a@b.c already exist.
      security: *a1
      summary: Create new administrator
      tags: *a2
  /administrators/{id}:
    get:
      operationId: AdministratorsController_getContactV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorDto"
        "404":
          description: Administrator requested not found
      security: *a1
      summary: Get administrator by ID
      tags: *a2
    delete:
      operationId: AdministratorsController_deleteAdministratorV1
      parameters:
        - name: id
          required: true
          in: path
          description: Administrator ID
          schema:
            type: number
      responses:
        "204":
          description: Administrator deleted successfully.
        "404":
          description: Administrator requested not found.
        "406":
          description: Administrator could not be deleted.
        "500":
          description: Administrator could not be deleted.
      security: *a1
      summary: Delete a administrator
      tags: *a2
    put:
      operationId: AdministratorsController_updateAdministratorV1
      parameters:
        - name: id
          required: true
          in: path
          description: Administrator ID
          schema:
            type: number
      requestBody:
        required: true
        description: Administrator DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministratorUpdateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorDto"
        "400":
          description: Bad request
          content:
            application/json:
              example: email must be an email
        "406":
          description: Administrator could not be update.
      security: *a1
      summary: Update an administrator
      tags: *a2
...
  /administrators-groups:
    get:
      operationId: AdministratorGroupsController_getContactsV1
      parameters:
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: nextPageToken
          required: false
          in: query
          schema:
            default: null
            type: string
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupsDto"
      security: &a14
        - bearer: []
      summary: Get list of administrator's groups
      tags: &a15
        - Administrators
    post:
      operationId: AdministratorGroupsController_createAdministratorV1
      parameters: []
      requestBody:
        required: true
        description: New administratorGroup DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministratorGroupCreateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupDto"
      security: *a14
      summary: Create new administratorGroup
      tags: *a15
  /administrators-groups/{id}:
    get:
      operationId: AdministratorGroupsController_getAdministratorGroupV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupDto"
        "404":
          description: Group requested not found
      security: *a14
      summary: Get administrator group by ID
      tags: *a15
    delete:
      operationId: AdministratorGroupsController_deleteAdministratorV1
      parameters:
        - name: id
          required: true
          in: path
          description: Administrators group ID
          schema:
            type: number
      responses:
        "204":
          description: Administrators group deleted successfully.
        "404":
          description: Administrators group requested not found.
        "406":
          description: Administrators group could not be deleted.
        "500":
          description: Administrators group could not be deleted.
      security: *a14
      summary: Delete a administrators group
      tags: *a15
    put:
      operationId: AdministratorGroupsController_updateAdministratorGroupV1
      parameters:
        - name: id
          required: true
          in: path
          description: Administrator group ID
          schema:
            type: number
      requestBody:
        required: true
        description: Administrator group DTO.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministratorGroupUpdateDto"
      responses:
        "200":
          description: Administrator group updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupDto"
        "406":
          description: Administrator group could not be update.
      security: *a14
      summary: Update an administrator group
      tags: *a15
    patch:
      operationId: AdministratorGroupsController_patchAdministratorGroupV1
      parameters:
        - name: id
          required: true
          in: path
          description: Administrator group ID
          schema:
            type: number
      requestBody:
        required: true
        description: Patch operations.
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PatchOperationDto"
      responses:
        "200":
          description: Administrator group updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministratorGroupDto"
        "406":
          description: Administrator group could not be update.
      security: *a14
      summary: Patch an administrator group
      tags: *a15
info:
  title: OnPage API
  description: >-
    OpenAPI definition for managing dispatching, pages, contacts,
    administrators, templates, settings, and webhooks.


    [📥 Download Swagger YAML](./swagger.yaml)
  version: "1.0"
  contact: {}
...
servers:
  - url: http://127.0.0.1:3003
    description: Development Server
  - url: https://public-api.onsetmobile.com
    description: Sandbox Server
  - url: https://public-api.onpage.com
    description: Production Server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AdministratorPermissions:
      type: object
      properties:
        createEscalation:
          type: boolean
          description: Can create escalation group
          default: false
        groupCreate:
          type: boolean
          description: Can create group
          default: false
        contactDelete:
          type: boolean
          description: Can delete contact
          default: false
        contactEdit:
          type: boolean
          description: Can edit contact
          default: false
        contactAdd:
          type: boolean
          description: Can create contact
          default: false
        contactToGroup:
          type: boolean
          description: Can add contact a group
          default: false
        removeContactFromGroup:
          type: boolean
          description: an remove from a group
          default: false
        deleteGroup:
          type: boolean
          description: Can delete group
          default: false
        editGroup:
          type: boolean
          description: Can edit group
          default: false
        editEscalationGroup:
          type: boolean
          description: Can edit escalation group
          default: false
        viewSchedule:
          type: boolean
          description: Can view scheduler
          default: false
        editSchedule:
          type: boolean
          description: Can edit scheduler
          default: false
        viewReports:
          type: boolean
          description: Can login to reports
          default: false
    AdministratorDto:
      type: object
      properties:
        id:
          type: number
          description: Administrator ID
        firstName:
          type: string
          description: Administrator first name
        lastName:
          type: string
          description: Administrator last name
        email:
          type: string
          description: Administrator email address
        phoneNumber:
          type: string
          description: Administrator phone number
        groups:
          description: Groups the administrator is member of
          type: array
          items:
            type: number
        superAdmin:
          type: boolean
          description: Super admin flag
          default: false
        permissions:
          description: Administrator permissions
          allOf:
            - $ref: "#/components/schemas/AdministratorPermissions"
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - permissions
    Metadata:
      type: object
      properties:
        nextPageToken:
          type: string
      required:
        - nextPageToken
    AdministratorsDto:
      type: object
      properties:
        administrators:
          description: Array of administrators
          type: array
          items:
            $ref: "#/components/schemas/AdministratorDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - administrators
        - metadata
    AdministratorCreateDto:
      type: object
      properties:
        password:
          type: string
          description: Administrator password
        firstName:
          type: string
          description: Administrator first name
        lastName:
          type: string
          description: Administrator last name
        email:
          type: string
          description: Administrator email address
        phoneNumber:
          type: string
          description: Administrator phone number
        groups:
          description: Groups the administrator is member of
          type: array
          items:
            type: number
        superAdmin:
          type: boolean
          description: Super admin flag
          default: false
        permissions:
          description: Administrator permissions
          allOf:
            - $ref: "#/components/schemas/AdministratorPermissions"
      required:
        - password
        - firstName
        - lastName
        - email
        - phoneNumber
        - permissions
    AdministratorUpdateDto:
      type: object
      properties:
        password:
          type: string
          description: Administrator password
        firstName:
          type: string
          description: Administrator first name
        lastName:
          type: string
          description: Administrator last name
        phoneNumber:
          type: string
          description: Administrator phone number
        groups:
          description: Groups the administrator is member of
          type: array
          items:
            type: number
        superAdmin:
          type: boolean
          description: Super admin flag
          default: false
        permissions:
          description: Administrator permissions
          allOf:
            - $ref: "#/components/schemas/AdministratorPermissions"
...
    PatchOperationDto:
      type: object
      properties:
        op:
          type: string
          description: The operation to be performed
          enum:
            - add
            - remove
            - replace
            - move
            - copy
            - test
        path:
          type: string
          description: A JSON Pointer that indicates the location to perform the operation
        value:
          type: object
          description: The value to be used for the operation (if applicable)
        from:
          type: string
          description: The source location for move or copy operations (if applicable)
      required:
        - op
        - path
...
    AdministratorGroupDto:
      type: object
      properties:
        id:
          type: number
          description: Administrator group ID
        name:
          type: string
          description: Administrator group name
        administrators:
          description: List of administrator IDs in the group
          type: array
          items:
            type: number
      required:
        - id
        - name
        - administrators
    AdministratorGroupsDto:
      type: object
      properties:
        groups:
          description: Administrator groups
          type: array
          items:
            $ref: "#/components/schemas/AdministratorGroupDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - groups
        - metadata
    AdministratorGroupCreateDto:
      type: object
      properties:
        name:
          type: string
          description: Administrator group name
        administrators:
          description: List of administrator IDs.
          type: array
          items:
            type: number
      required:
        - name
    AdministratorGroupUpdateDto:
      type: object
      properties:
        name:
          type: string
          description: Administrator group name
        administrators:
          description: List of administrator IDs.
          type: array
          items:
            type: number
      required:
        - name